Test: move functional test to separate file
